requirements.txt
numpy==1.22.2
pandas==1.0.0
Pillow==8.4.0
tkcalendar==1.6.1
xlrd==1.2.0
XlsxWriter==3.0.2
